Break data service integration test into separate files
